+++ view/frontend/requirejs-config.js
@@ -1,0 +1,7 @@
+var config = {
+    map: {
+        '*': {
+            'fotorama/fotorama':'Scaleflex_Cloudimage/js/fotorama'
+        }
+    }
+};