+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.2] - 2022-04-08
+
+* Add Fotorama Compatibility Mode option
+* Add Device Pixel Max Ratio option
+
 ## [2.0.1] - 2022-02-01
 
 * Remove "/v7" from the URL part option added